delete done

+++ app/Http/Controllers/PengelolaController.php
@@ -54,7 +54,7 @@ class PengelolaController extends Controller
                 '<nobr>' . 
                     '<a href="'.route(self::URL .'show', $row->id).'" class="btn btn-success btn-sm" title="Detail"><i class="fas fa-folder"></i> Detail</a> ' .
                     '<a href="'.route(self::URL .'edit', $row->id).'" class="btn btn-info btn-sm" title="Edit"><i class="fas fa-pencil-alt"></i> Edit</a> ' .
-                    // '<button type="button" class="btn btn-danger btn-sm btnDelete" value="'.$row->id.'" id="'.route(self::URL . 'destroy', $row->id).'"><i class="fas fa-trash"></i> Hapus</button>' . 
+                    '<button type="button" class="btn btn-danger btn-sm btnDelete" value="'.$row->id.'" id="'.route(self::URL . 'destroy', $row->id).'"><i class="fas fa-trash"></i> Hapus</button>' . 
                 '</nobr>',
             ]);
 
@@ -183,9 +183,19 @@ class PengelolaController extends Controller
             return abort(403, "User does not have the right roles");
         }
 
-        $row->delete();
+        $kontaks = $row->pengelola_kontaks->count();
+        $dokumens = $row->pengelola_dokumens->count();
 
-        return response()->json('Success');
+        if (
+            empty($kontaks) &&
+            empty($dokumens)
+        ) {
+            $row->delete();
+
+            return response()->json('Success');
+        } else {
+            return response()->json('Data tidak bisa di hapus, karena sudah mempunyai hubungan dibawahnya.', 403);
+        }
     }
 
     public function apiList(Request $request)
